third_party/WebKit/LayoutTests/fast/html/imports/import-custom-element-abort-expected.txt

CONSOLE MESSAGE: line 4: PASS unless crash.
